+++ app/src/main/java/com/example/weatherapplication/domain/model/day/DayImperialTemperature.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapplication.domain.model.day
 
 data class DayImperialTemperature(
-    override val maxTemp: Double,
-    override val minTemp: Double,
-    override val avgTemp: Double
+    override val maxTemp: Int,
+    override val minTemp: Int,
+    override val avgTemp: Int
 ) : DayTemperature
